SPARKC-4: Refactoring: Simplify ColumnRef hierarchy and make ColumnMappers check any mapping errors.
Remove ColumnIndex. Column references are now always by-name.
This allowed to move most of the properties of intermediate traits like
NamedColumnRef or SelectableColumnRef to the base class,
to flatten the hierarchy and then to remove pattern matching
on ColumnRef in many places.

Without ColumnIndex, ColumnMappers now must know the column names
to return proper named ColumnRefs in the ColumnMaps.
Therefore they are given full column selection information in a
form of a list of ColumnRefs selected by the user. Some ColumnMappers
as TupleColumnMapper, which need to reference column by their index
in the selection list, simply look up appropriate ColumnRefs when creating
the result ColumnMap. This simplifies further processing like row reading,
because values are  fetched always by-name, never by the index.
It is also less-likely to mess up the column order.

Knowing the list of selected columns, ColumnMappers can validate
the mapping, for example detect missing columns early.
Because validation is different for reading and for writing,
ColumnMap has been split into ColumnMapForReading and ColumnMapForWriting,
which are produced by two separate methods of ColumnMapper.
For writing, the mapped class must provide getters for all selected columns.
For reading, the selected column set must contain at least all columns
that are to be used when calling object's constructor.

Previously ColumnMappers didn't do any validation and were allowed
to create invalid ColumnMaps. They just
mapped properties to "hypothetical" columns that hopefully would be
present in the table definition, but they never checked if they were
*really* present and *selected by the user*.
The code using the ColumnMapper needed to validate the created
ColumnMap afterwards. Due to the fact ColumnMappers are used in many places,
it was easy to miss some error checks and errors would be reported late,
probably when executing the Spark task, where the information on what
has caused that was limited.