SPARKC-4: Replace cqlValueName with selectedAs in CassandraRDDSpec
When defining custom column mapping, we should use the public alias
of the column determined by the column selector, not the internal name
used in the CQL result set. In this case, they were fortunately the
same, so the test did pass. But in principle, cqlValueName
is not something the user code should rely on.